Improve serach recommendations dag
