--- app/airflow/dags/libs/search_recommendations/__init__.py
@@ -1,1 +1,0 @@
-# Search recommendations module for Airflow DAGs
